+++ src/main/java/zupzup/back_end/reservation/Exception/OrderRuntimeException.java
@@ -1,0 +1,15 @@
+package zupzup.back_end.reservation.Exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class OrderRuntimeException extends RuntimeException {
+
+    private final HttpStatus httpStatus;
+
+    public OrderRuntimeException(final String message, final HttpStatus httpStatus) {
+        super(message);
+        this.httpStatus = httpStatus;
+    }
+}
